+++ packages/apollos-data-connector-postgres/src/people/migrations/index.js
@@ -1,0 +1,11 @@
+import CreatePeople001 from './001_create_people';
+import ChangeProfileImageType002 from './002_change_profile_image_url_to_text';
+import AddNamesIndex003 from './003_add_names_index';
+
+const migrations = [
+  CreatePeople001,
+  ChangeProfileImageType002,
+  AddNamesIndex003,
+];
+
+export default migrations;