+++ packages/apollos-data-schema/src/index.js
@@ -189,8 +189,23 @@ export const peopleSchema = gql`
     uploadProfileImage(file: Upload!, size: Int!): Person
   }
 
+  type SearchPeopleResultsConnection {
+    edges: [SearchPeopleResult]
+    pageInfo: PaginationInfo
+  }
+
+  type SearchPeopleResult {
+    node: Person
+    cursor: String
+  }
+
   extend type Query {
     suggestedFollows: [Person] @cacheControl(maxAge: 0)
+    searchPeople(
+      name: String
+      first: Int
+      after: String
+    ): SearchPeopleResultsConnection
   }
 `;
 
@@ -1037,6 +1052,10 @@ export const commentSchema = gql`
     ): Comment
 
     flagComment(commentId: ID!): Comment
+
+    likeComment(commentId: ID!): Comment
+
+    unlikeComment(commentId: ID!): Comment
   }
 
   type CommentListFeature implements Feature & Node {
@@ -1061,12 +1080,14 @@ export const commentSchema = gql`
     FOLLOWERS
   }
 
-  type Comment implements Node {
+  type Comment implements Node & LikableNode {
     id: ID!
 
     person: Person
     text: String
     visibility: CommentVisibility
+    isLiked: Boolean @cacheControl(maxAge: 0)
+    likedCount: Int
   }
 `;