+++ packages/apollos-ui-auth/src/ProfileDetails/ProfileDetailsEntry.js
@@ -25,9 +25,9 @@ const ProfileDetailsEntry = (props) => (
     >
       {props.genderList.map((gender) => [
         <RadioButton
-          key={gender}
-          value={gender}
-          label={() => <RadioLabel>{gender}</RadioLabel>}
+          key={gender.field || gender}
+          value={gender.field || gender}
+          label={() => <RadioLabel>{gender.label || gender}</RadioLabel>}
           underline={false}
         />,
       ])}
@@ -80,7 +80,15 @@ ProfileDetailsEntry.propTypes = {
   }),
   BackgroundComponent: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
   onPressBack: PropTypes.func.isRequired,
-  genderList: PropTypes.arrayOf(PropTypes.string),
+  genderList: PropTypes.arrayOf(
+    PropTypes.oneOfType([
+      PropTypes.string,
+      PropTypes.shape({
+        field: PropTypes.string,
+        label: PropTypes.string,
+      }),
+    ])
+  ),
   defaultDate: PropTypes.instanceOf(Date),
 };
 
@@ -88,7 +96,11 @@ ProfileDetailsEntry.defaultProps = {
   title: "This one's easy.",
   prompt:
     'Help us understand who you are so we can connect you with the best ministries and events.',
-  genderList: ['Male', 'Female', 'Prefer not to reply'],
+  genderList: [
+    'Male',
+    'Female',
+    { label: 'Prefer not to reply', field: 'Unknown' },
+  ],
 };
 
 ProfileDetailsEntry.LegalText = LegalText;