+++ packages/apollos-ui-connected/src/CommentListFeatureConnected/CommentListFeatureConnected.js
@@ -6,6 +6,8 @@ import { get } from 'lodash';
 import PropTypes from 'prop-types';
 import GET_COMMENT_LIST_FEATURE from './getCommentListFeature';
 import FLAG_COMMENT from './flagComment';
+import LIKE_COMMENT from './likeComment';
+import UNLIKE_COMMENT from './unlikeComment';
 import CommentListFeature from './CommentListFeature';
 
 const presentActionOption = ({ callback, actionText, title }) => {
@@ -50,7 +52,48 @@ function CommentListFeatureConnected({
     fetchPolicy: 'cache-and-network',
   });
 
-  const [flagComment] = useMutation(FLAG_COMMENT);
+  const onFlagComment = (cache, { data: { flagComment } }) => {
+    // Let's get started! No crashing allowed.
+    try {
+      // Let's find our comment list.
+      const commentListFeature = cache.readQuery({
+        query: GET_COMMENT_LIST_FEATURE,
+        variables: { featureId },
+      });
+
+      // Now let's remove our flagged comment from that comment list.
+      cache.writeQuery({
+        query: GET_COMMENT_LIST_FEATURE,
+        variables: { featureId },
+        data: {
+          ...commentListFeature,
+          node: {
+            ...commentListFeature.node,
+            comments: commentListFeature.node.comments.filter(
+              ({ id }) => id !== flagComment.id
+            ),
+          },
+        },
+      });
+    } catch (e) {
+      console.warn('Failed to update cache after adding comment', e);
+    }
+  };
+
+  const [flagComment] = useMutation(FLAG_COMMENT, {
+    update: onFlagComment,
+  });
+
+  const [likeComment] = useMutation(LIKE_COMMENT);
+  const [unlikeComment] = useMutation(UNLIKE_COMMENT);
+
+  const handlePressLike = ({ isLiked, id }) => {
+    if (isLiked) {
+      unlikeComment({ variables: { commentId: id } });
+    } else {
+      likeComment({ variables: { commentId: id } });
+    }
+  };
 
   const handlePressActionMenu = ({ id: commentId }) => {
     presentActionOption({
@@ -69,7 +112,7 @@ function CommentListFeatureConnected({
       {...props}
       isLoading={loading || isLoading}
       onPressActionMenu={handlePressActionMenu}
-      onFlagComment
+      onPressLike={handlePressLike}
     />
   );
 }