+++ packages/apollos-ui-notifications/src/NotificationSettingsConnected/NotificationSettings.tests.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+
+import { Providers } from '@apollosproject/ui-kit';
+
+import NotificationSettings from './NotificationSettings';
+
+describe('The NotificationSettings toggle switch', () => {
+  it('should render correctly', () => {
+    const tree = renderer.create(
+      <Providers>
+        <NotificationSettings />
+      </Providers>
+    );
+    expect(tree).toMatchSnapshot();
+  });
+});