+++ packages/apollos-data-connector-postgres/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.0](https://github.com/apollosproject/apollos-apps/compare/v2.7.0...v2.8.0) (2021-03-29)
+
+
+### Bug Fixes
+
+* Map gender to a postgres compatible value ([#1836](https://github.com/apollosproject/apollos-apps/issues/1836)) ([183da8b](https://github.com/apollosproject/apollos-apps/commit/183da8bfafa7f810d77cd470f8f74ae8dcd688e2))
+
+
+
+## [2.7.0](https://github.com/apollosproject/apollos-apps/compare/v2.6.0...v2.7.0) (2021-03-25)
+
+
+### Features
+
+* Liking comments ([#1829](https://github.com/apollosproject/apollos-apps/issues/1829)) ([4590af7](https://github.com/apollosproject/apollos-apps/commit/4590af7c4299d3e7e0643b25385d82cfc7891ada))
+
+
+
+## [2.6.0](https://github.com/apollosproject/apollos-apps/compare/v2.5.0...v2.6.0) (2021-03-24)
+
+
+### Bug Fixes
+
+* Don't search non-apollosUsers and trim whitespace ([#1820](https://github.com/apollosproject/apollos-apps/issues/1820)) ([3599340](https://github.com/apollosproject/apollos-apps/commit/35993404922e8d9ae3bdfe275a6f2488f1ae2c8b))
+
+
+
+## [2.5.0](https://github.com/apollosproject/apollos-apps/compare/v2.4.0...v2.5.0) (2021-03-22)
+
+
+### Features
+
+* add bottom sheet to AddCommentForm ([#1807](https://github.com/apollosproject/apollos-apps/issues/1807)) ([4c09ff0](https://github.com/apollosproject/apollos-apps/commit/4c09ff0382c60aa011080407de5a6443d5b35db5))
+* Auto follow ([#1810](https://github.com/apollosproject/apollos-apps/issues/1810)) ([a96bfec](https://github.com/apollosproject/apollos-apps/commit/a96bfec991ba1f4f444183d40b525b89063a31e9))
+* People Search ([#1815](https://github.com/apollosproject/apollos-apps/issues/1815)) ([16e3556](https://github.com/apollosproject/apollos-apps/commit/16e35562198d9b0049e067075c53f9f9aab0d25f))
+* Push notifications on follow request ([#1806](https://github.com/apollosproject/apollos-apps/issues/1806)) ([19ba81b](https://github.com/apollosproject/apollos-apps/commit/19ba81b0452c19275918fbdc251425fa981a735e))
+
+
+
+## [2.4.0](https://github.com/apollosproject/apollos-apps/compare/v2.3.1...v2.4.0) (2021-03-18)
+
+
+### Features
+
+* Adds Postgres migrations ([#1791](https://github.com/apollosproject/apollos-apps/issues/1791)) ([eb115b6](https://github.com/apollosproject/apollos-apps/commit/eb115b616eaa6f1a69c5c0701bbb8dd99e0b4dbe))
+* Adds RequestedFollowsList  ([#1792](https://github.com/apollosproject/apollos-apps/issues/1792)) ([b855156](https://github.com/apollosproject/apollos-apps/commit/b85515670375785deb366c787819020656aee29c))
+* Sort users you follow to the top ([#1801](https://github.com/apollosproject/apollos-apps/issues/1801)) ([d6360a4](https://github.com/apollosproject/apollos-apps/commit/d6360a49fa60d34b0dc1f870b16a05458b8e428a))
+
+
+### Bug Fixes
+
+* Upgrades and fixes for comments ([#1800](https://github.com/apollosproject/apollos-apps/issues/1800)) ([b77dbc2](https://github.com/apollosproject/apollos-apps/commit/b77dbc28af621190173e7c82e0e4295abc047109))
+
+
+
 ### [2.3.1](https://github.com/apollosproject/apollos-apps/compare/v2.3.0...v2.3.1) (2021-03-11)
 
 **Note:** Version bump only for package @apollosproject/data-connector-postgres